refactor: Restructure project folder, handle null values, switch to RequiredArgsConstructor, method renaming, error handling improvements
This commit includes the following changes:
- Restructured the project folder to improve organization
- Implemented null value handling in relevant parts of the code
- Replaced @Autowired with @RequiredArgsConstructor for constructor injection
- Renamed methods to improve clarity and consistency
- Enhanced error handling across the project for better reliability

+++ src/main/java/cursos/alura/api/domain/rating/RatingCourseNpsResultDTO.java
@@ -12,9 +12,9 @@ import lombok.Setter;
 public class RatingCourseNpsResultDTO {
     private String codeCourse;
     private String courseName;
-    private int promoters;
-    private int neutrals;
-    private int detractors;
+    private long promoters;
+    private long neutrals;
+    private long detractors;
     private boolean hasEnoughRatings;
     private String nps;
